pyproject.toml
[project]
name = "users-also-buy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "basedpyright>=1.34.0",
    "gql[requests]>=3.5.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.26.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.7",
    "typesense>=1.3.0",
    "streamlit>=1.40.1",
    "openpyxl>=3.1.5",
    "gql[requests]>=4.0.0",
]
